requirements.txt
opencv-python
numpy
tqdm
paddlepaddle -i https://mirror.baidu.com/pypi/simple
paddleocr>=2.6.0.3
paddleclas>=2.4.3
# will update soon
pdf2image
paddleocr
paddlepaddle
#paddlepaddle-gpu
docxcompose
premailer
openpyxl

transformers
sentencepiece
#cpu
torch==2.0.1 
torchvision==0.15.2 
torchaudio==2.0.2 
#sacremoses.
pyspellchecker